Save and user messages context for private chat with bot

+++ src/main/scala/undiy/aibot/context/ContextService.scala
@@ -5,6 +5,6 @@ import undiy.aibot.context.model.ContextMessage
 trait ContextService[F[_]] {
   def saveContextMessage(message: ContextMessage): F[Unit]
 
-  def getContextMessages(chatId: Int, limit: Int = 100): F[List[ContextMessage]]
+  def getContextMessages(chatId: Long, limit: Int = 100): F[List[ContextMessage]]
 }
